// src/domain/models/product/product.ts
export interface ProductModel {
  name: string;
  description: string;
  price: number;
  stock: number;
  category: string;
  status: boolean;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  stock: number;
  category: string;
  status: boolean;
}